内存安全三层接 CI：ASan+UBSan（sanitizers job）、cppcheck+clang-tidy（static-analysis job）、编译期加固（RA2R_HARDEN：栈保护/_GLIBCXX_ASSERTIONS/_FORTIFY_SOURCE + -Wconversion -Wshadow 零告警）；修复静态分析存量告警（AUD ADPCM 越界截断、object_layer art 指针重构、Obj/Sec/Blowfish 默认初始化、lcw 条件自增、vxl 名称截断等）

+++ engine/include/ra2r/assets/fnt_file.h
@@ -63,7 +63,7 @@ public:
     // TS 变体：字符 ch（0..255 内，超出或越界返回空）
     FntGlyph ts_glyph(int ch) const {
         FntGlyph g;
-        if (kind_ != FntKind::Ts || ch < 0 || ch >= ts_char_count()) return g;
+        if (kind_ != FntKind::Ts || ch < 0 || ch >= static_cast<int>(ts_char_count())) return g;
         g.width = ts_cx_[ch];
         g.data = data_.data() + ts_img_ofs_[ch];
         g.stride_bytes = (ts_cx_[ch] + 7) / 8; // 行宽按像素取整
